SOINNへパスつないだ． これからは作業フォルダにsoinnパッケージがないとエラーになるはずなので注意
行列のベクトル化とベクトルの行列化のメソッドを作成．

+++ IntentionLearning_v2/src/komota/lib/MyIO.java
@@ -197,7 +197,7 @@ public class MyIO {
 			}
 			String[] splitted = line.split(",");
 			for(int idx=0;idx<dim;idx++){
-				read[count][idx] = Integer.parseInt(splitted[idx].replace("	", ""));
+				read[count][idx] = Double.parseDouble(splitted[idx].replace("	", ""));
 			}
 			count++;
 		}
